backend/django-monolith/app/views.py
from rest_framework import generics
from .models import Asset
from .serializers import AssetSerializer

class AssetListCreateView(generics.ListCreateAPIView):
    queryset = Asset.objects.all()
    serializer_class = AssetSerializer 